ocr: init
copy text on screen to the clipboard, how convenient.
still, there is a bug with warframe (the only for now) that prevents me
from ocr it. overlay should be above fullscreen, right?
solution: screenshot it and open on viewnior, ocr the chat you want and
translate with anyrun.
this is dumb i know, but is temporary (i.e. it will become a feature if
i forget about it)