ags: fix notification sizing
It was an annoying for a long time, and it's 'fixed' without any nasty
hacks.
This introduced another problem, though. The entire notification column
responds to clicks, making everything below that layer not clickable
until the layer closes.
This commit also adds a viewport for notifications that are too long, to
avoid filling up the screen when someone sends you an entire essay (it
happened before)