ags: fix notification window size
Bottom anchor was making the notifications 'window' take a good chunk of
the screen.
Widget now cuts everything after 140 characters into a scrollable
viewport, making it possible for notifications to have a standard size
Viewport size is a bit small for now, and I'd prefer if the limit was
higher. But it's the only way if we don't want to use bottom anchor.