audio/pipewire: low latency 'refactor'
It's still a copy, but if it works it works.
For some reason, using "context.properties" syntax makes PipeWire shit
itself on startup, and restarting the service after that makes the audio
a distorted mess.
Also, this commit adds some missing udev rules in audio/realtime and a
enables the "threadirqs" kernel option