modules: add system.nixos.tags to roles and profiles
We still can't reliably infer the type of the host using this.
Taking `polaris` for example, it has tags `desktop` and `server`, and
you may think that's contradictory.
It is.
Using this, we can enable some options, but to reliably get the host
type we should be defining it under `nice.host`, something like
`nice.host.type`